[AST] Introduce MangledNameTypeBuilder.
And change RemoteASTTypeBuilder to use it. We plan to
use TypeDecoder in TypeReconstruction and this is the first
step toward the goal as it abstract type creation that doesn't
require AST lookups.
